refactor: kakao api 관련 상수 관리 클래스 이름 변경
